switch from int to long

+++ src/header.hpp
@@ -12,7 +12,7 @@ typedef struct s_Node Node;
 struct s_Node{
 	Node * left_son;
 	Node * right_son;
-	int value;
+	long value;
 };
 
 //Die verschiedenen Varianten des Speicher Layouts.
@@ -22,13 +22,13 @@ enum Layout {RANDOM, SORTED, LAYERS, EMDEBOAS};
 //Funktionen
 //
 
-//Generiert einen Baum mit 'size' Einträgen und 'mem_type' Speicherlayout.
+//Generiert einen Baum mit 'size' Elongrägen und 'mem_type' Speicherlayout.
 //
-Node* generate_tree(unsigned int size, enum Layout mem_type, unsigned int* numbers, int blockheight);
+Node* generate_tree(unsigned long size, enum Layout mem_type, unsigned long* numbers, long blockheight);
 
 //Suche im 'tree' nach dem Schlüssel 'key'
 //@return: die Suchzeit in Millisekunden zurück
 // ODER -1, wenn der Schlüssel nicht gefunden wird.
-bool search_in_tree(Node* tree, unsigned int key);
+bool search_in_tree(Node* tree, unsigned long key);
 
 #endif
